Move Funnel from numberAxis folder to cartesian folder
so that it matches the API description and is next to other similar files
