+++ src/cartesian/Funnel.tsx
