+++ src/index.ts
@@ -99,8 +99,8 @@ export { RadialBarChart } from './chart/RadialBarChart';
 export { ComposedChart } from './chart/ComposedChart';
 export { SunburstChart } from './chart/SunburstChart';
 
-export { Funnel } from './numberAxis/Funnel';
-export type { FunnelProps } from './numberAxis/Funnel';
+export { Funnel } from './cartesian/Funnel';
+export type { FunnelProps } from './cartesian/Funnel';
 export { FunnelChart } from './chart/FunnelChart';
 export { Trapezoid } from './shape/Trapezoid';
 export type { Props as TrapezoidProps } from './shape/Trapezoid';
